move creation of AWSRequest
Still throwing AWSClient Errors though
