+++ Sources/AWSSDKSwiftCore/AWSClient.swift
@@ -24,9 +24,8 @@ import class  Foundation.JSONSerialization
 import class  Foundation.JSONDecoder
 import struct Foundation.Data
 import struct Foundation.URL
-import struct Foundation.URLComponents
 import struct Foundation.URLQueryItem
-import struct Foundation.CharacterSet
+
 
 /// This is the workhorse of aws-sdk-swift-core. You provide it with a `AWSShape` Input object, it converts it to `AWSRequest` which is then converted
 /// to a raw `HTTPClient` Request. This is then sent to AWS. When the response from AWS is received if it is successful it is converted to a `AWSResponse`
@@ -422,227 +421,6 @@ extension AWSClient {
             return AWSSigner(credentials: credential, name: self.serviceConfig.signingName, region: self.serviceConfig.region.rawValue)
         }
     }
-    
-}
-
-// request creation
-extension AWSRequest {
-    
-    internal static func verifyStream(operation: String, payload: AWSPayload, input: AWSShapeWithPayload.Type) {
-        guard case .stream(let reader) = payload.payload else { return }
-        precondition(input._payloadOptions.contains(.allowStreaming), "\(operation) does not allow streaming of data")
-        precondition(reader.size != nil || input._payloadOptions.contains(.allowChunkedStreaming), "\(operation) does not allow chunked streaming of data. Please supply a data size.")
-    }
-    
-    internal init<Input: AWSEncodableShape>(
-        operation operationName: String,
-        path: String,
-        httpMethod: String,
-        input: Input,
-        configuration: ServiceConfig) throws
-    {
-        var headers: [String: Any] = [:]
-        var path = path
-        var body: Body = .empty
-        var queryParams: [(key:String, value:Any)] = []
-
-        // validate input parameters
-        try input.validate()
-
-        guard let baseURL = URL(string: "\(configuration.endpoint)"), let _ = baseURL.host else {
-            throw AWSClient.ClientError.invalidURL("\(configuration.endpoint) must specify url host and scheme")
-        }
-
-        // set x-amz-target header
-        if let target = configuration.amzTarget {
-            headers["x-amz-target"] = "\(target).\(operationName)"
-        }
-
-        // TODO should replace with Encodable
-        let mirror = Mirror(reflecting: input)
-        var memberVariablesCount = mirror.children.count - Input._encoding.count
-
-        // extract header, query and uri params
-        for encoding in Input._encoding {
-            if let value = mirror.getAttribute(forKey: encoding.label) {
-                switch encoding.location {
-                case .header(let location):
-                    headers[location] = value
-
-                case .querystring(let location):
-                    switch value {
-                    case let array as QueryEncodableArray:
-                        array.queryEncoded.forEach { queryParams.append((key:location, value:$0)) }
-                    case let dictionary as QueryEncodableDictionary:
-                        dictionary.queryEncoded.forEach { queryParams.append($0) }
-                    default:
-                        queryParams.append((key:location, value:"\(value)"))
-                    }
-
-                case .uri(let location):
-                    path = path
-                        .replacingOccurrences(of: "{\(location)}", with: "\(value)")
-                        // percent-encode key which is part of the path
-                        .replacingOccurrences(of: "{\(location)+}", with: "\(value)".addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)!)
-
-                default:
-                    memberVariablesCount += 1
-                }
-            }
-        }
-
-        switch configuration.serviceProtocol {
-        case .json, .restjson:
-            if let shapeWithPayload = Input.self as? AWSShapeWithPayload.Type {
-                let payload = shapeWithPayload._payloadPath
-                if let payloadBody = mirror.getAttribute(forKey: payload) {
-                    switch payloadBody {
-                    case let awsPayload as AWSPayload:
-                        Self.verifyStream(operation: operationName, payload: awsPayload, input: shapeWithPayload)
-                        body = .raw(awsPayload)
-                    case let shape as AWSEncodableShape:
-                        body = .json(try shape.encodeAsJSON())
-                    default:
-                        preconditionFailure("Cannot add this as a payload")
-                    }
-                } else {
-                    body = .empty
-                }
-            } else {
-                // only include the body if there are members that are output in the body.
-                if memberVariablesCount > 0 {
-                    body = .json(try input.encodeAsJSON())
-                }
-            }
-
-        case .query:
-            var dict = try input.encodeAsQuery()
-
-            dict["Action"] = operationName
-            dict["Version"] = configuration.apiVersion
-
-            switch httpMethod {
-            case "GET":
-                queryParams.append(contentsOf: dict.map {(key:$0.key, value:$0)})
-            default:
-                if let urlEncodedQueryParams = Self.urlEncodeQueryParams(fromDictionary: dict) {
-                    body = .text(urlEncodedQueryParams)
-                }
-            }
-
-        case .restxml:
-            if let shapeWithPayload = Input.self as? AWSShapeWithPayload.Type {
-                let payload = shapeWithPayload._payloadPath
-                if let payloadBody = mirror.getAttribute(forKey: payload) {
-                    switch payloadBody {
-                    case let awsPayload as AWSPayload:
-                        Self.verifyStream(operation: operationName, payload: awsPayload, input: shapeWithPayload)
-                        body = .raw(awsPayload)
-                    case let shape as AWSEncodableShape:
-                        var rootName: String? = nil
-                        // extract custom payload name
-                        if let encoding = Input.getEncoding(for: payload), case .body(let locationName) = encoding.location {
-                            rootName = locationName
-                        }
-                        body = .xml(try shape.encodeAsXML(rootName: rootName))
-                    default:
-                        preconditionFailure("Cannot add this as a payload")
-                    }
-                } else {
-                    body = .empty
-                }
-            } else {
-                // only include the body if there are members that are output in the body.
-                if memberVariablesCount > 0 {
-                    body = .xml(try input.encodeAsXML())
-                }
-            }
-
-        case .ec2:
-            var params = try input.encodeAsQueryForEC2()
-            params["Action"] = operationName
-            params["Version"] = configuration.apiVersion
-            if let urlEncodedQueryParams = Self.urlEncodeQueryParams(fromDictionary: params) {
-                body = .text(urlEncodedQueryParams)
-            }
-        }
-
-        guard let parsedPath = URLComponents(string: path) else {
-            throw AWSClient.ClientError.invalidURL("\(configuration.endpoint)\(path)")
-        }
-
-        // add queries from the parsed path to the query params list
-        if let pathQueryItems = parsedPath.queryItems {
-            for item in pathQueryItems {
-                queryParams.append((key:item.name, value: item.value ?? ""))
-            }
-        }
-
-        // Build URL. Don't use URLComponents as Foundation and AWS disagree on what should be percent encoded in the query values
-        var urlString = "\(baseURL.absoluteString)\(parsedPath.path)"
-        if queryParams.count > 0 {
-            urlString.append("?")
-            urlString.append(queryParams.sorted{$0.key < $1.key}.map{"\($0.key)=\(Self.urlEncodeQueryParam("\($0.value)"))"}.joined(separator:"&"))
-        }
-
-        guard let url = URL(string: urlString) else {
-            throw AWSClient.ClientError.invalidURL("\(urlString)")
-        }
-        
-        self.region = configuration.region
-        self.url = url
-        self.serviceProtocol = configuration.serviceProtocol
-        self.operation = operationName
-        self.httpMethod = httpMethod
-        self.httpHeaders = headers
-        self.body = body
-
-//        .applyMiddlewares(serviceConfig.middlewares + middlewares)
-    }
-    
-    // this list of query allowed characters comes from https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html
-    static let queryAllowedCharacters = CharacterSet(charactersIn:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~")
-
-    private static func urlEncodeQueryParam(_ value: String) -> String {
-        return value.addingPercentEncoding(withAllowedCharacters: AWSRequest.queryAllowedCharacters) ?? value
-    }
-
-    private static func urlEncodeQueryParams(fromDictionary dict: [String:Any]) -> String? {
-        guard dict.count > 0 else {return nil}
-        var query = ""
-        let keys = Array(dict.keys).sorted()
-
-        for iterator in keys.enumerated() {
-            let value = dict[iterator.element]
-            query += iterator.element + "=" + (urlEncodeQueryParam(String(describing: value ?? "")))
-            if iterator.offset < dict.count - 1 {
-                query += "&"
-            }
-        }
-        return query
-    }
-    
-    internal init(operation operationName: String, path: String, httpMethod: String, configuration: ServiceConfig) throws {
-        var headers: [String: Any] = [:]
-
-        guard let url = URL(string: "\(configuration.endpoint)\(path)"), let _ = url.host else {
-            throw AWSClient.ClientError.invalidURL("\(configuration.endpoint)\(path) must specify url host and scheme")
-        }
-
-        // set x-amz-target header
-        if let target = configuration.amzTarget {
-            headers["x-amz-target"] = "\(target).\(operationName)"
-        }
-        
-        self.region = configuration.region
-        self.url = url
-        self.serviceProtocol = configuration.serviceProtocol
-        self.operation = operationName
-        self.httpMethod = httpMethod
-        self.httpHeaders = headers
-        self.body = .empty
-    }
-    
     
 }
